pipeline/pyproject.toml
[project]
name = "efta-pipeline"
version = "0.1.0"
description = "Epstein Files Tracking & Analysis - Document Processing Pipeline"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-storage>=2.18.0",
    "google-cloud-firestore>=2.19.0",
    "google-cloud-documentai>=2.32.0",
    "google-cloud-pubsub>=2.23.0",
    "google-cloud-secret-manager>=2.21.0",
    "google-cloud-aiplatform>=1.72.0",
    "google-cloud-speech>=2.28.0",
    "google-genai>=1.0.0",
    "pydantic>=2.9.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "cryptography>=43.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
